Added .upipelinecache support
Not too sure how useful it will be, but I could implement its use as a cache (you can't 100% depend on your shader information being there)

+++ src/lib/Objects/Core/EPixelFormat.h
@@ -3,7 +3,7 @@
 #include <string>
 
 namespace upp::Objects {
-    enum class EPixelFormat : uint8_t {
+    enum class EPixelFormat : uint32_t {
         PF_Unknown = 0,
         PF_A32B32G32R32F = 1,
         PF_B8G8R8A8 = 2,
